requirements.txt
ipykernel
pandas
scikit-learn
numpy
matplotlib
seaborn
streamlit
cassandra-driver